UPDATE: New get methods for users module
- Get user by shop_id
- Get operator by user_id
- Get merchant by shop_id

+++ src/modules/users/services/operators.service.ts
@@ -132,6 +132,47 @@ const getByShopId = async ({ shop_id }: Pick<IOperator, "shop_id">) => {
   };
 };
 
+const getByUserId = async ({ user_id }: Pick<IOperator, "user_id">) => {
+  try {
+    const result = await Operator.findAll({
+      attributes: [
+        [sequelize.literal('BIN_TO_UUID(Operator.id)'), 'id'],
+        [sequelize.literal('BIN_TO_UUID(Operator.user_id)'), 'user_id'],
+        [sequelize.literal('BIN_TO_UUID(Operator.shop_id)'), 'shop_id'],
+        'created_at',
+        'updated_at'
+      ],
+      include: [
+        {
+          model: User,
+          as: 'user',
+          attributes: [
+            [sequelize.literal('BIN_TO_UUID(user.id)'), 'id'],
+            'full_name',
+            'email',
+            'created_at',
+            'updated_at',
+          ],
+        },
+        {
+          model: Shop,
+          as: 'shop',
+          attributes: [
+            [sequelize.literal('BIN_TO_UUID(shop.id)'), 'id'],
+            'name'
+          ],
+        },
+      ],
+      where: sequelize.literal(`user.id = UUID_TO_BIN(?)`),
+      replacements: [user_id],
+    });
+
+    return result.length ? result : null;
+  } catch (error) {
+    throw new Error("Error al obtener el operador por el id de usuario.");
+  }
+};
+
 const getByUserAndShop = async ({ user_id, shop_id }: Pick<IOperator, "user_id" | "shop_id">) => {
   try {
     const result = await Operator.findAll({
@@ -198,4 +239,4 @@ const deleteById = async ({ id }: Pick<IOperator, "id">) => {
 };
 
 
-export default { add, getAll, getById, getByShopId, deleteById, getByUserAndShop };
+export default { add, getAll, getById, getByShopId, getByUserId, deleteById, getByUserAndShop };
